digest uses godirwalk
