+++ internal/gps/pkgtree/digest.go
@@ -12,112 +12,20 @@ import (
 	"io"
 	"os"
 	"path/filepath"
+	"sort"
 	"strconv"
 
+	"github.com/karrick/godirwalk"
 	"github.com/pkg/errors"
 )
 
-const osPathSeparator = string(filepath.Separator)
+const (
+	bufsize               = 16 * 1024
+	osPathSeparatorLength = 1
+)
 
-// lineEndingReader is a `io.Reader` that converts CRLF sequences to LF.
-//
-// When cloning or checking out repositories, some Version Control Systems,
-// VCSs, on some supported Go Operating System architectures, GOOS, will
-// automatically convert line endings that end in a single line feed byte, LF,
-// to line endings that end in a two byte sequence of carriage return, CR,
-// followed by LF. This LF to CRLF conversion would cause otherwise identical
-// versioned files to have different on disk contents simply based on which VCS
-// and GOOS are involved. Different file contents for the same file would cause
-// the resultant hashes to differ. In order to ensure file contents normalize
-// and produce the same hash, this structure wraps an io.Reader that modifies
-// the file's contents when it is read, translating all CRLF sequences to LF.
-type lineEndingReader struct {
-	src             io.Reader // source io.Reader from which this reads
-	prevReadEndedCR bool      // used to track whether final byte of previous Read was CR
-}
-
-// newLineEndingReader returns a new lineEndingReader that reads from the
-// specified source io.Reader.
-func newLineEndingReader(src io.Reader) *lineEndingReader {
-	return &lineEndingReader{src: src}
-}
-
-var crlf = []byte("\r\n")
-
-// Read consumes bytes from the structure's source io.Reader to fill the
-// specified slice of bytes. It converts all CRLF byte sequences to LF, and
-// handles cases where CR and LF straddle across two Read operations.
-func (f *lineEndingReader) Read(buf []byte) (int, error) {
-	buflen := len(buf)
-	if f.prevReadEndedCR {
-		// Read one fewer bytes so we have room if the first byte of the
-		// upcoming Read is not a LF, in which case we will need to insert
-		// trailing CR from previous read.
-		buflen--
-	}
-	nr, er := f.src.Read(buf[:buflen])
-	if nr > 0 {
-		if f.prevReadEndedCR && buf[0] != '\n' {
-			// Having a CRLF split across two Read operations is rare, so the
-			// performance impact of copying entire buffer to the right by one
-			// byte, while suboptimal, will at least will not happen very
-			// often. This negative performance impact is mitigated somewhat on
-			// many Go compilation architectures, GOARCH, because the `copy`
-			// builtin uses a machine opcode for performing the memory copy on
-			// possibly overlapping regions of memory. This machine opcodes is
-			// not instantaneous and does require multiple CPU cycles to
-			// complete, but is significantly faster than the application
-			// looping through bytes.
-			copy(buf[1:nr+1], buf[:nr]) // shift data to right one byte
-			buf[0] = '\r'               // insert the previous skipped CR byte at start of buf
-			nr++                        // pretend we read one more byte
-		}
-
-		// Remove any CRLF sequences in the buffer using `bytes.Index` because,
-		// like the `copy` builtin on many GOARCHs, it also takes advantage of a
-		// machine opcode to search for byte patterns.
-		var searchOffset int // index within buffer from whence the search will commence for each loop; set to the index of the end of the previous loop.
-		var shiftCount int   // each subsequenct shift operation needs to shift bytes to the left by one more position than the shift that preceded it.
-		previousIndex := -1  // index of previously found CRLF; -1 means no previous index
-		for {
-			index := bytes.Index(buf[searchOffset:nr], crlf)
-			if index == -1 {
-				break
-			}
-			index += searchOffset // convert relative index to absolute
-			if previousIndex != -1 {
-				// shift substring between previous index and this index
-				copy(buf[previousIndex-shiftCount:], buf[previousIndex+1:index])
-				shiftCount++ // next shift needs to be 1 byte to the left
-			}
-			previousIndex = index
-			searchOffset = index + 2 // start next search after len(crlf)
-		}
-		if previousIndex != -1 {
-			// handle final shift
-			copy(buf[previousIndex-shiftCount:], buf[previousIndex+1:nr])
-			shiftCount++
-		}
-		nr -= shiftCount // shorten byte read count by number of shifts executed
-
-		// When final byte from a read operation is CR, do not emit it until
-		// ensure first byte on next read is not LF.
-		if f.prevReadEndedCR = buf[nr-1] == '\r'; f.prevReadEndedCR {
-			nr-- // pretend byte was never read from source
-		}
-	} else if f.prevReadEndedCR {
-		// Reading from source returned nothing, but this struct is sitting on a
-		// trailing CR from previous Read, so let's give it to client now.
-		buf[0] = '\r'
-		nr = 1
-		er = nil
-		f.prevReadEndedCR = false // prevent infinite loop
-	}
-	return nr, er
-}
-
-// writeBytesWithNull appends the specified data to the specified hash, followed by
-// the NULL byte, in order to make accidental hash collisions less likely.
+// writeBytesWithNull appends the specified data to the specified hash, followed
+// by the NULL byte, in order to make accidental hash collisions less likely.
 func writeBytesWithNull(h hash.Hash, data []byte) {
 	// Ignore return values from writing to the hash, because hash write always
 	// returns nil error.
@@ -127,10 +35,11 @@ func writeBytesWithNull(h hash.Hash, data []byte) {
 // dirWalkClosure is used to reduce number of allocation involved in closing
 // over these variables.
 type dirWalkClosure struct {
-	someCopyBufer []byte // allocate once and reuse for each file copy
-	someModeBytes []byte // allocate once and reuse for each node
-	someDirLen    int
-	someHash      hash.Hash
+	someCopyBufer     []byte // allocate once and reuse for each file copy
+	someModeBytes     []byte // allocate once and reuse for each node
+	someScratchBuffer []byte // allocate once and reuse for every directory read
+	someDirLen        int
+	someHash          hash.Hash
 }
 
 // DigestFromDirectory returns a hash of the specified directory contents, which
@@ -150,110 +59,94 @@ type dirWalkClosure struct {
 //
 // While filepath.Walk could have been used, that standard library function
 // skips symbolic links, and for now, we want the hash to include the symbolic
-// link referents.
+// link referents. In addition, the filepath.Walk function invokes os.Stat for
+// each file system node discovered, while the godirwalk.WalkFileMode fetches
+// the file system node type while it's reading the parent directory.
 func DigestFromDirectory(osDirname string) ([]byte, error) {
+	return digestFromDirectoryBuffer(osDirname, make([]byte, bufsize))
+}
+
+func digestFromDirectoryBuffer(osDirname string, scratchBuffer []byte) ([]byte, error) {
 	osDirname = filepath.Clean(osDirname)
 
 	// Create a single hash instance for the entire operation, rather than a new
 	// hash for each node we encounter.
 
 	closure := dirWalkClosure{
-		someCopyBufer: make([]byte, 4*1024), // only allocate a single page
-		someModeBytes: make([]byte, 4),      // scratch place to store encoded os.FileMode (uint32)
-		someDirLen:    len(osDirname) + len(osPathSeparator),
-		someHash:      sha256.New(),
+		someCopyBufer:     make([]byte, 4*1024), // only allocate a single page
+		someDirLen:        len(osDirname) + osPathSeparatorLength,
+		someHash:          sha256.New(),
+		someModeBytes:     make([]byte, 4), // scratch place to store encoded os.FileMode (uint32)
+		someScratchBuffer: scratchBuffer,
 	}
 
-	err := DirWalk(osDirname, func(osPathname string, info os.FileInfo, err error) error {
-		if err != nil {
-			return err // DirWalk received an error during initial Lstat
-		}
-
-		var osRelative string
-		if len(osPathname) > closure.someDirLen {
-			osRelative = osPathname[closure.someDirLen:]
-		}
-
-		switch filepath.Base(osRelative) {
-		case "vendor", ".bzr", ".git", ".hg", ".svn":
-			return filepath.SkipDir
-		}
-
-		// We could make our own enum-like data type for encoding the file type,
-		// but Go's runtime already gives us architecture independent file
-		// modes, as discussed in `os/types.go`:
-		//
-		//    Go's runtime FileMode type has same definition on all systems, so
-		//    that information about files can be moved from one system to
-		//    another portably.
-		var mt os.FileMode
-
-		// We only care about the bits that identify the type of a file system
-		// node, and can ignore append, exclusive, temporary, setuid, setgid,
-		// permission bits, and sticky bits, which are coincident to bits which
-		// declare type of the file system node.
-		modeType := info.Mode() & os.ModeType
-		var shouldSkip bool // skip some types of file system nodes
-
-		switch {
-		case modeType&os.ModeDir > 0:
-			mt = os.ModeDir
-			// DirWalkFunc itself does not need to enumerate children, because
-			// DirWalk will do that for us.
-			shouldSkip = true
-		case modeType&os.ModeSymlink > 0:
-			mt = os.ModeSymlink
-		case modeType&os.ModeNamedPipe > 0:
-			mt = os.ModeNamedPipe
-			shouldSkip = true
-		case modeType&os.ModeSocket > 0:
-			mt = os.ModeSocket
-			shouldSkip = true
-		case modeType&os.ModeDevice > 0:
-			mt = os.ModeDevice
-			shouldSkip = true
-		}
-
-		// Write the relative pathname to hash because the hash is a function of
-		// the node names, node types, and node contents. Added benefit is that
-		// empty directories, named pipes, sockets, devices, and symbolic links
-		// will also affect final hash value. Use `filepath.ToSlash` to ensure
-		// relative pathname is os-agnostic.
-		writeBytesWithNull(closure.someHash, []byte(filepath.ToSlash(osRelative)))
-
-		binary.LittleEndian.PutUint32(closure.someModeBytes, uint32(mt)) // encode the type of mode
-		writeBytesWithNull(closure.someHash, closure.someModeBytes)      // and write to hash
-
-		if shouldSkip {
-			return nil // nothing more to do for some of the node types
-		}
-
-		if mt == os.ModeSymlink { // okay to check for equivalence because we set to this value
-			osRelative, err = os.Readlink(osPathname) // read the symlink referent
-			if err != nil {
-				return errors.Wrap(err, "cannot Readlink")
+	err := godirwalk.Walk(osDirname, &godirwalk.Options{
+		FollowSymbolicLinks: true,
+		Callback: func(osPathname string, de *godirwalk.Dirent) error {
+			if de.IsDir() {
+				switch de.Name() {
+				case "vendor", ".bzr", ".git", ".hg", ".svn":
+					return filepath.SkipDir
+				}
 			}
-			writeBytesWithNull(closure.someHash, []byte(filepath.ToSlash(osRelative))) // write referent to hash
-			return nil                                                                 // proceed to next node in queue
-		}
 
-		// If we get here, node is a regular file.
-		fh, err := os.Open(osPathname)
-		if err != nil {
-			return errors.Wrap(err, "cannot Open")
-		}
+			var err error
+			var osRelative string // os-specific pathname with osDirname prefix removed
 
-		var bytesWritten int64
-		bytesWritten, err = io.CopyBuffer(closure.someHash, newLineEndingReader(fh), closure.someCopyBufer) // fast copy of file contents to hash
-		err = errors.Wrap(err, "cannot Copy")                                                               // errors.Wrap only wraps non-nil, so skip extra check
-		writeBytesWithNull(closure.someHash, []byte(strconv.FormatInt(bytesWritten, 10)))                   // 10: format file size as base 10 integer
+			if len(osPathname) > closure.someDirLen {
+				osRelative = osPathname[closure.someDirLen:]
+			}
 
-		// Close the file handle to the open file without masking
-		// possible previous error value.
-		if er := fh.Close(); err == nil {
-			err = errors.Wrap(er, "cannot Close")
-		}
-		return err
+			// Write the relative pathname to hash because the hash is a function of
+			// the node names, node types, and node contents. Added benefit is that
+			// empty directories, named pipes, sockets, devices, and symbolic links
+			// will also affect final hash value. Use `filepath.ToSlash` to ensure
+			// relative pathname is os-agnostic.
+			writeBytesWithNull(closure.someHash, []byte(filepath.ToSlash(osRelative)))
+
+			modeType := de.ModeType()
+			binary.LittleEndian.PutUint32(closure.someModeBytes, uint32(modeType)) // encode the type of mode
+			writeBytesWithNull(closure.someHash, closure.someModeBytes)            // and write to hash
+
+			switch {
+			case modeType&os.ModeSymlink != 0:
+				osRelative, err = os.Readlink(osPathname) // read the symlink referent
+				if err != nil {
+					return errors.Wrap(err, "cannot Readlink")
+				}
+				writeBytesWithNull(closure.someHash, []byte(filepath.ToSlash(osRelative))) // write referent to hash
+				return nil                                                                 // proceed to next node
+			case modeType&os.ModeDir != 0:
+				return nil // nothing more to do for this type
+			case modeType&os.ModeDevice != 0:
+				return nil // nothing more to do for this type
+			case modeType&os.ModeCharDevice != 0:
+				return nil // nothing more to do for this type
+			case modeType&os.ModeNamedPipe != 0:
+				return nil // nothing more to do for this type
+			case modeType&os.ModeSocket != 0:
+				return nil // nothing more to do for this type
+			}
+
+			// If we get here, node is a regular file.
+			fh, err := os.Open(osPathname)
+			if err != nil {
+				return errors.Wrap(err, "cannot Open")
+			}
+
+			var bytesWritten int64
+			bytesWritten, err = io.CopyBuffer(closure.someHash, newLineEndingReader(fh), closure.someCopyBufer) // fast copy of file contents to hash
+			err = errors.Wrap(err, "cannot Copy")                                                               // errors.Wrap only wraps non-nil, so skip extra check
+			writeBytesWithNull(closure.someHash, []byte(strconv.FormatInt(bytesWritten, 10)))                   // 10: format file size as base 10 integer
+
+			// Close the file handle to the open file without masking possible
+			// previous error value.
+			if er := fh.Close(); err == nil {
+				err = errors.Wrap(er, "cannot Close")
+			}
+			return err
+		},
+		ScratchBuffer: closure.someScratchBuffer,
 	})
 	if err != nil {
 		return nil, err
@@ -391,6 +284,9 @@ func VerifyDepTree(osDirname string, wantSums map[string][]byte) (map[string]Ven
 		slashStatus[slashPathname] = NotInTree
 	}
 
+	// create a scratch buffer for raw bytes from reading directory entries
+	scratchBuffer := make([]byte, bufsize)
+
 	for len(queue) > 0 {
 		// Pop node from the top of queue (depth first traversal, reverse
 		// lexicographical order inside a directory), clearing the value stored
@@ -403,7 +299,7 @@ func VerifyDepTree(osDirname string, wantSums map[string][]byte) (map[string]Ven
 		if expectedSum, ok := wantSums[slashPathname]; ok {
 			ls := EmptyDigestInLock
 			if len(expectedSum) > 0 {
-				projectSum, err := DigestFromDirectory(osPathname)
+				projectSum, err := digestFromDirectoryBuffer(osPathname, scratchBuffer)
 				if err != nil {
 					return nil, errors.Wrap(err, "cannot compute dependency hash")
 				}
@@ -425,28 +321,29 @@ func VerifyDepTree(osDirname string, wantSums map[string][]byte) (map[string]Ven
 			continue
 		}
 
-		osChildrenNames, err := sortedChildrenFromDirname(osPathname)
+		deChildren, err := godirwalk.ReadDirents(osPathname, scratchBuffer)
 		if err != nil {
-			return nil, errors.Wrap(err, "cannot get sorted list of directory children")
+			return nil, errors.Wrap(err, "cannot get list of directory children")
 		}
-		for _, osChildName := range osChildrenNames {
-			switch osChildName {
-			case ".", "..", "vendor", ".bzr", ".git", ".hg", ".svn":
+		sort.Sort(deChildren)
+		for _, deChild := range deChildren {
+			name := deChild.Name()
+			switch name {
+			case "vendor", ".bzr", ".git", ".hg", ".svn":
 				// skip
 			default:
-				osChildRelative := filepath.Join(currentNode.osRelative, osChildName)
-				osChildPathname := filepath.Join(osDirname, osChildRelative)
+				osChildRelative := filepath.Join(currentNode.osRelative, name)
 
 				// Create a new fsnode for this file system node, with a parent
 				// index set to the index of the current node.
-				otherNode := &fsnode{osRelative: osChildRelative, myIndex: len(nodes), parentIndex: currentNode.myIndex}
-
-				fi, err := os.Stat(osChildPathname)
-				if err != nil {
-					return nil, errors.Wrap(err, "cannot Stat")
+				otherNode := &fsnode{
+					osRelative:  osChildRelative,
+					myIndex:     len(nodes),
+					parentIndex: currentNode.myIndex,
 				}
+
 				nodes = append(nodes, otherNode) // Track all file system nodes...
-				if fi.IsDir() {
+				if deChild.IsDir() {
 					queue = append(queue, otherNode) // but only need to add directories to the work queue.
 				}
 			}
