+++ internal/gps/pkgtree/digest_test.go
@@ -6,96 +6,15 @@ package pkgtree
 
 import (
 	"bytes"
-	"io"
 	"os"
 	"path/filepath"
 	"testing"
 )
 
-// crossBuffer is a test io.Reader that emits a few canned responses.
-type crossBuffer struct {
-	readCount  int
-	iterations []string
-}
-
-func (cb *crossBuffer) Read(buf []byte) (int, error) {
-	if cb.readCount == len(cb.iterations) {
-		return 0, io.EOF
-	}
-	cb.readCount++
-	return copy(buf, cb.iterations[cb.readCount-1]), nil
-}
-
-func streamThruLineEndingReader(t *testing.T, iterations []string) []byte {
-	dst := new(bytes.Buffer)
-	n, err := io.Copy(dst, newLineEndingReader(&crossBuffer{iterations: iterations}))
-	if got, want := err, error(nil); got != want {
-		t.Errorf("(GOT): %v; (WNT): %v", got, want)
-	}
-	if got, want := n, int64(dst.Len()); got != want {
-		t.Errorf("(GOT): %v; (WNT): %v", got, want)
-	}
-	return dst.Bytes()
-}
-
-func TestLineEndingReader(t *testing.T) {
-	testCases := []struct {
-		input  []string
-		output string
-	}{
-		{[]string{"\r"}, "\r"},
-		{[]string{"\r\n"}, "\n"},
-		{[]string{"now is the time\r\n"}, "now is the time\n"},
-		{[]string{"now is the time\r\n(trailing data)"}, "now is the time\n(trailing data)"},
-		{[]string{"now is the time\n"}, "now is the time\n"},
-		{[]string{"now is the time\r"}, "now is the time\r"},     // trailing CR ought to convey
-		{[]string{"\rnow is the time"}, "\rnow is the time"},     // CR not followed by LF ought to convey
-		{[]string{"\rnow is the time\r"}, "\rnow is the time\r"}, // CR not followed by LF ought to convey
-
-		// no line splits
-		{[]string{"first", "second", "third"}, "firstsecondthird"},
-
-		// 1->2 and 2->3 both break across a CRLF
-		{[]string{"first\r", "\nsecond\r", "\nthird"}, "first\nsecond\nthird"},
-
-		// 1->2 breaks across CRLF and 2->3 does not
-		{[]string{"first\r", "\nsecond", "third"}, "first\nsecondthird"},
-
-		// 1->2 breaks across CRLF and 2 ends in CR but 3 does not begin LF
-		{[]string{"first\r", "\nsecond\r", "third"}, "first\nsecond\rthird"},
-
-		// 1 ends in CR but 2 does not begin LF, and 2->3 breaks across CRLF
-		{[]string{"first\r", "second\r", "\nthird"}, "first\rsecond\nthird"},
-
-		// 1 ends in CR but 2 does not begin LF, and 2->3 does not break across CRLF
-		{[]string{"first\r", "second\r", "\nthird"}, "first\rsecond\nthird"},
-
-		// 1->2 and 2->3 both break across a CRLF, but 3->4 does not
-		{[]string{"first\r", "\nsecond\r", "\nthird\r", "fourth"}, "first\nsecond\nthird\rfourth"},
-		{[]string{"first\r", "\nsecond\r", "\nthird\n", "fourth"}, "first\nsecond\nthird\nfourth"},
-
-		{[]string{"this is the result\r\nfrom the first read\r", "\nthis is the result\r\nfrom the second read\r"},
-			"this is the result\nfrom the first read\nthis is the result\nfrom the second read\r"},
-		{[]string{"now is the time\r\nfor all good engineers\r\nto improve their test coverage!\r\n"},
-			"now is the time\nfor all good engineers\nto improve their test coverage!\n"},
-		{[]string{"now is the time\r\nfor all good engineers\r", "\nto improve their test coverage!\r\n"},
-			"now is the time\nfor all good engineers\nto improve their test coverage!\n"},
-	}
-
-	for _, testCase := range testCases {
-		got := streamThruLineEndingReader(t, testCase.input)
-		if want := []byte(testCase.output); !bytes.Equal(got, want) {
-			t.Errorf("Input: %#v; (GOT): %#q; (WNT): %#q", testCase.input, got, want)
-		}
-	}
-}
-
-////////////////////////////////////////
-
-func getTestdataVerifyRoot(t *testing.T) string {
+func getTestdataVerifyRoot(tb testing.TB) string {
 	cwd, err := os.Getwd()
 	if err != nil {
-		t.Fatal(err)
+		tb.Fatal(err)
 	}
 	return filepath.Join(filepath.Dir(cwd), "testdata_digest")
 }
@@ -137,6 +56,32 @@ func TestDigestFromDirectory(t *testing.T) {
 			t.Errorf("\n(GOT):\n\t%#v\n(WNT):\n\t%#v", got, want)
 		}
 	})
+}
+
+func BenchmarkDigestFromDirectory(b *testing.B) {
+	prefix := getTestdataVerifyRoot(b)
+	b.ResetTimer()
+
+	for i := 0; i < b.N; i++ {
+		_, err := DigestFromDirectory(prefix)
+		if err != nil {
+			b.Fatal(err)
+		}
+	}
+}
+
+const flameIterations = 100000
+
+func BenchmarkFlameDigestFromDirectory(b *testing.B) {
+	prefix := getTestdataVerifyRoot(b)
+	b.ResetTimer()
+
+	for i := 0; i < flameIterations; i++ {
+		_, err := DigestFromDirectory(prefix)
+		if err != nil {
+			b.Fatal(err)
+		}
+	}
 }
 
 func TestVerifyDepTree(t *testing.T) {
@@ -193,23 +138,8 @@ func TestVerifyDepTree(t *testing.T) {
 	}
 }
 
-func BenchmarkDigestFromDirectory(b *testing.B) {
-	b.Skip("Eliding benchmark of user's Go source directory")
-
-	prefix := filepath.Join(os.Getenv("GOPATH"), "src")
-
-	for i := 0; i < b.N; i++ {
-		_, err := DigestFromDirectory(prefix)
-		if err != nil {
-			b.Fatal(err)
-		}
-	}
-}
-
 func BenchmarkVerifyDepTree(b *testing.B) {
-	b.Skip("Eliding benchmark of user's Go source directory")
-
-	prefix := filepath.Join(os.Getenv("GOPATH"), "src")
+	prefix := getTestdataVerifyRoot(b)
 
 	for i := 0; i < b.N; i++ {
 		_, err := VerifyDepTree(prefix, nil)
